Move ack_bytes to the one place it is used
